backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
psutil==5.9.6
websockets==12.0
python-multipart==0.0.6
